+++ BankApplication/Classes/User.cs
@@ -28,7 +28,7 @@ namespace BankApplication
         [ForeignKey("Clients")]
         public int Id_User { get; set; }
         public Clients Clients { get; set; }
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
         [Key]
         public string NumberAccount { get; set; }
         public decimal Debet { get; set; }
@@ -54,7 +54,7 @@ namespace BankApplication
         public Clients Clients { get; set; }
         public string NumberAccountSender { get; set; }
         public string NumberAccountRecipient { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Title { get; set; }
         [Key]
         public int Id_Transfer { get; set; }
